Правки

+++ tests/web/conftest.py
@@ -17,46 +17,49 @@ def load_env():
 
 @pytest.fixture(scope="function", autouse=True)
 def setup_browser(request):
-    browser.config.window_width = 1920
-    browser.config.window_height = 1080
-    browser.config.base_url = "https://www.fatsecret.com"
+    if "skip_setup_browser" in request.keywords:
+        yield
+    else:
+        browser.config.window_width = 1920
+        browser.config.window_height = 1080
+        browser.config.base_url = "https://www.fatsecret.com"
 
-    options = Options()
-    selenoid_capabilities = {
-        "browserName": "chrome",
-        "browserVersion": "126.0",
-        "selenoid:options": {"enableVNC": True, "enableVideo": True},
-    }
+        options = Options()
+        selenoid_capabilities = {
+            "browserName": "chrome",
+            "browserVersion": "126.0",
+            "selenoid:options": {"enableVNC": True, "enableVideo": True},
+        }
 
-    options.add_argument("--disable-gpu")
-    options.add_argument("--no-sandbox")
-    options.add_argument("--disable-dev-shm-usage")
-    options.add_argument("--incognito")
+        options.add_argument("--disable-gpu")
+        options.add_argument("--no-sandbox")
+        options.add_argument("--disable-dev-shm-usage")
+        options.add_argument("--incognito")
 
-    selenoid_login = os.getenv("SELENOID_LOGIN")
-    selenoid_pass = os.getenv("SELENOID_PASS")
-    selenoid_url = os.getenv("SELENOID_URL")
+        selenoid_login = os.getenv("SELENOID_LOGIN")
+        selenoid_pass = os.getenv("SELENOID_PASS")
+        selenoid_url = os.getenv("SELENOID_URL")
 
-    options.capabilities.update(selenoid_capabilities)
-    driver = webdriver.Remote(
-        command_executor=f"https://{selenoid_login}:{selenoid_pass}@{selenoid_url}/wd/hub",
-        options=options,
-    )
+        options.capabilities.update(selenoid_capabilities)
+        driver = webdriver.Remote(
+            command_executor=f"https://{selenoid_login}:{selenoid_pass}@{selenoid_url}/wd/hub",
+            options=options,
+        )
 
-    browser.config.driver = driver
+        browser.config.driver = driver
 
-    yield
+        yield
 
-    attach.add_screenshot()
-    attach.add_xml()
-    attach.add_video_web()
-    attach.add_logs()
+        attach.add_screenshot()
+        attach.add_xml()
+        attach.add_video_web()
+        attach.add_logs()
 
-    if "test_authorization" not in request.fspath.basename:
-        add_food = AddFood()
-        delete_food = DeleteFood()
+        if "test_authorization" not in request.fspath.basename:
+            add_food = AddFood()
+            delete_food = DeleteFood()
 
-        add_food.go_to_food_diary()
-        delete_food.clear_diary()
+            add_food.go_to_food_diary()
+            delete_food.clear_diary()
 
-    browser.quit()
+        browser.quit()
